+++ src/table/parse/index.js
@@ -27,7 +27,9 @@ const cleanText = function (str) {
   str = str.replace(/style=['"].*?["']/, '')
   //'!' is used as a highlighed-column
   str = str.replace(/^!/, '')
+  str = str.replace(/\(.*?\)/, '')
   str = str.trim()
+  str = str.toLowerCase()
   return str
 }